move tests into their own files added permission checks for investment tasks
